requirements.txt
# Flask and extensions
flask==3.0.0
flask-cors==4.0.0

# HTTP requests for APIs
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# CORS and Werkzeug
Werkzeug==3.0.1
